Add node discovery and message sending to MeshNetwork
Added new method to the MeshNetwork struct to discover nodes and send messages. Also made changes to mesh and device errors to improve error handling. Plus, the 'snafu' crate was added to the Cargo.toml for simplified error handling. Route structure was updated in the network route file to change next_hop's type to Uid. Modifications in the message file include the creation of a Message Error struct for more specific error handling and adjustment in the message structure fields. Lastly, the device file was revised to handle device-specific errors. These modifications enhance the network's routing and device messaging specialties, which increases the overall efficiency.