Replace hard-coded message size with constant
This commit replaces the hard-coded values of the message size with a constant. These changes make code more maintainable and minimize errors - specifically, if the message size needs to be changed in the future, it now only requires a single change. The potential range issue is resolved by adjusting indexing on byte arrays.