Add receive_message function to MeshNetwork
A new function `receive_message` is added to the `MeshNetwork` struct. The function is responsible for receiving messages from the network, deserializing them into a `Message` object, updating the routing table based on the sender of the message, and checking whether the received message is intended for the current node or needs to be forwarded to the next hop. This change is necessary to handle incoming messages and update network routing in mesh networks.