Add structure for handling intent in messages
This update introduces an intent structure within the Message system to allow for cleaner handling of diverse message types. The Intent enum includes Data, Discover, Ping, Pong, Information, and Error types. Attempting to convert a byte not represented in the Intent enum will result in an InvalidIntent error.

Changes also made to other structs to accommodate this addition, and several methods updated from using u16 to u8.

Specific constructors for each message type were also introduced to make the message creation process more explicit and less prone to errors.

Some placeholder sections and version-specific information yet to be implemented, particularly for the Information intent handling.

(Notably, the message size was reduced from 74 to 70 bytes)

These changes will improve the organization and readability of the code and is a step towards a more efficient and robust communication system.